requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==43.0.1
distlib==0.3.6
dulwich==0.21.3
Faker==28.4.1
filelock==3.9.0
html5lib==1.1
idna==3.7
importlib-metadata==6.0.0
iniconfig==2.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.5.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.3.3
python-dateutil==2.9.0.post0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
tomlkit==0.11.6
trove-classifiers==2023.3.9
urllib3==2.0.7
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
zipp==3.15.0
